Starting point. Several tests are written, everything compiles, tests all pass. We are in a great place from which to start... or are we...
It's really easy to go back after the fact and write some tests and not cover every contingency. I think I did, but there may be some paths through the conditionals that are not well tested.

Also, some people write a test, run it to get the generated output, and encode that output as the expected values without really understanding where the numbers come from.

Testing first is so much better... trust me!