Starting down path of improving names of variables and methods in system. name to customerName is the first one, and several will be bundled into the next commit. The idea behind this is two-fold: Improving the names requires reading and understand the code, and changing the names encodes that learning in the source code itself.
One point of caution - I changed the name of a public method, which may have larger impacts outside of this class. It may be better to add a new method, deprecate the old one, and remove it over time. I was able to make that leap, you may not be.