Completed renaming variables. My strategy is to make the variable name explain what the variable is for, irrespective of the type of the variable. This strategy may change a bit if I'm in a dynamic language and I want to encode a little type-knowledge into the nanme. I tend not to do this, but I have before as a reminder.
I also try to make similar things be named the same, and use adjectives to make variables that refer to similar concepts be easily recognizable as related. See totalRentalAmount and totalFrequentRenterPoints, and currentRental and currentRentalAmount. Some redundancy in the naming, but it makes it obvious which things go which which concepts.