Noticed that I didn't really need the currentRentalAmount temporary variable, as it was only caching a simply calculated value.
Big question is what to do if the calculation isn't quick, simple, and cheap? Do you not do the refactoring in that case? That would be setting up a false dichotomy of either do the refactoring or not. There are other options, like caching the result of the calculation, that would allow the refactoring to stil happen to improve the code. If the call is not idempotent, however, and you want to reuse the same value for the entire time through the loop, you may have to cache it at the top of your code.

All refactorings like this are contextually dependent. Use your judgement when making them.