Just split the for-loop into three separate loops. This is the step that frequently generates the most conversation... You can do this if the steps inside of the loops are independent of each other, as they are in our example. If they are related, then this refactoring may require you to chunk together different groupings of lines inside of a loop, or just conclude that the refactoring isn't possible for you.
The other question that comes up is what happens if iterating over the loop is expensive, like if you have an array of a million items or retrieving the items is expensive. This becomes a judgement call. Readability matters as a proprerty of code, so you have to judge if the performance penalty of iterating over the same loop three times is worth it. You could put the iteration into a method and cache the result so you only have to retrieve the entries once, as an example of how to work around the issue.