So the two methods newly moved into Rental use data from both the rental class and the movie class. I believe they need to be moved into the Movie class because the dependency on internals from the Movie class inside of the methods is a worse smell in the Rental class than needing the daysRented inside of each method if they were moved into Movie.
In the former case, the Rental is reaching inside of the Movie class and pulling out data then doing something with it. This is a violation of the Law of Demeter (look it up!). We can still get the results we need by moving the methods into the Movie class and just passing the daysRented into each method. This gives the Rental control of its own internal data and does not add to the dependencies of the solution.

Both methods moved into Movie, and delegating methods left in Rental. Now this brings up the question about whether to inline those two delegating methods. You can argue that the methods don't add much value and should be inlined in their caller (Customer). The other side of that argument is that the benefit these methods add is that they increase the encapsulation of the Rental class. If we inlined these methods, then any caller of them would have go do the same work each of them do internally - get the movie from the rental, get the rentalDays from the rental, then call the methods on the returned movie. These internal details are hidden by keeping these methods, so I believe they are very much worth keeping.

Next is to look at the methods in the Movie class and figure out how to deal with that ugly type switching logic.