OK, this was a big change. We had the switch statement in the Movie class, which is really bad. Anytime you switch on a type to alter behavior, you are missing a chance to use polymorphism. So we are going to apply the Replace Conditional with Polymorphism (RCP) refactoring. This uses several other refactorings to get its job done, like Self Encapsulate Field (SEF) and Replace Typecode with State/Strategy (RTSS) (look these up, too).
In this refactoring, we started applying RCP by creating an inheritance hierarchy where the state or strategy pattern will eventually live. We do this by creating an abstract base class, then one overriding subclass for each different value the state can have (this is the Movie typecode that we're looking at). This led me to create the MovieType, ChildrensMovieType, NewReleaseMovieType, and RegularMovieType. I gave each of the knowledge of their typecode, which they can return using polymorphism when asked.

Then I applied the SEF refactoring in Movie to make sure that all references to priceCode (need to change names to match here! priceCode and movieType are both referring to the same thing but with two name. That's our next refactoring) pass through a getter and setter. Doing this lets me change behavior throughout the entire class by only changing the implementation of the getter and setter.

So I first made sure that all calls inside of Movie when through getPriceCode and setPriceCode, then changed the behavior of those two methods. The getter now uses the polymorphism created in the new class hierarchy, and the setter creates an instance of the correct movie type based on the price code. This is an OK use of a switch statement, to create an object based on some type code.

I'm going to proceed with the refactoring for now, as I'm still noodling over whether the two different names (priceCode and movieType) are OK. They just night be...