Finished! We changed statement() into a method of my favorite length, 1 line!
I'll go through the code one more time cleaning up anything I missed and check that in as the last commit.

Thanks for reading along. I enjoy doing this example because it gives me a chance to take a 40+ line method and show how refactoring can pull out the algorithms and concepts hidihng in the code. I hope it was useful to you all!